Move struct parsing to a separate module
